Use lambda invoke function for checksum verification
Previously we updated the ttl to an expired date,
which AWS would pick up on to trigger a DDB event
that would invoke the lambda. However, this is
documented to be non-deterministic and there can
be a 48 hr lag between ttl expiry and the event
trigger firing. Not good for tests. So instead of
fully simulating that e2e just provide the event
payload and invoke the function.