Update tech docs
Heavily streamlined. LLM was used for docs because was difficult
to keep up with rapid iterations. We're still in an implemention
mode so more changes are to be expected. Better to keep things
minimal for now, stick to the essentials and consider how to
approach in the future when things are more properly dialled in.